requirements.txt
fastapi
uvicorn
clerk-backend-api
requests
python-dotenv
